resovle PR comments 3rd round
