+++ test/spark/test_multiclass.py
@@ -5,7 +5,7 @@ from sklearn.datasets import load_iris, load_wine
 from flaml import AutoML
 from flaml.automl.data import CLASSIFICATION, get_output_from_log
 from flaml.automl.training_log import training_log_reader
-from flaml.spark.utils import check_spark, customize_learner
+from flaml.spark.utils import check_spark
 import os
 
 try:
@@ -17,134 +17,30 @@ except (ImportError, RuntimeError):
 
 os.environ["FLAML_MAX_CONCURRENT"] = "2"
 
-learner_code = """
-from flaml import tune
-from flaml.automl.model import LGBMEstimator, XGBoostSklearnEstimator, SKLearnEstimator
-from flaml.automl.data import CLASSIFICATION, get_output_from_log
+# To solve pylint issue, we put code for customizing mylearner in a separate file
+if os.path.exists(os.path.join(os.getcwd(), "test", "spark", "custom_mylearner.py")):
+    try:
+        from test.spark.custom_mylearner import *
+        from flaml.spark.mylearner import (
+            MyRegularizedGreedyForest,
+            custom_metric,
+            MyLargeLGBM,
+            MyLargeXGB,
+        )
 
-class MyRegularizedGreedyForest(SKLearnEstimator):
-    def __init__(self, task="binary", **config):
-
-        super().__init__(task, **config)
-
-        if task in CLASSIFICATION:
-            from rgf.sklearn import RGFClassifier
-
-            self.estimator_class = RGFClassifier
-        else:
-            from rgf.sklearn import RGFRegressor
-
-            self.estimator_class = RGFRegressor
-
-    @classmethod
-    def search_space(cls, data_size, task):
-        space = {
-            "max_leaf": {
-                "domain": tune.lograndint(lower=4, upper=data_size[0]),
-                "init_value": 4,
-            },
-            "n_iter": {
-                "domain": tune.lograndint(lower=1, upper=data_size[0]),
-                "init_value": 1,
-            },
-            "n_tree_search": {
-                "domain": tune.lograndint(lower=1, upper=32768),
-                "init_value": 1,
-            },
-            "opt_interval": {
-                "domain": tune.lograndint(lower=1, upper=10000),
-                "init_value": 100,
-            },
-            "learning_rate": {"domain": tune.loguniform(lower=0.01, upper=20.0)},
-            "min_samples_leaf": {
-                "domain": tune.lograndint(lower=1, upper=20),
-                "init_value": 20,
-            },
-        }
-        return space
-
-    @classmethod
-    def size(cls, config):
-        max_leaves = int(round(config.get("max_leaf", 1)))
-        n_estimators = int(round(config.get("n_iter", 1)))
-        return (max_leaves * 3 + (max_leaves - 1) * 4 + 1.0) * n_estimators * 8
-
-    @classmethod
-    def cost_relative2lgbm(cls):
-        return 1.0
-
-
-class MyLargeXGB(XGBoostSklearnEstimator):
-    @classmethod
-    def search_space(cls, **params):
-        return {
-            "n_estimators": {
-                "domain": tune.lograndint(lower=4, upper=32768),
-                "init_value": 32768,
-                "low_cost_init_value": 4,
-            },
-            "max_leaves": {
-                "domain": tune.lograndint(lower=4, upper=3276),
-                "init_value": 3276,
-                "low_cost_init_value": 4,
-            },
-        }
-
-
-class MyLargeLGBM(LGBMEstimator):
-    @classmethod
-    def search_space(cls, **params):
-        return {
-            "n_estimators": {
-                "domain": tune.lograndint(lower=4, upper=32768),
-                "init_value": 32768,
-                "low_cost_init_value": 4,
-            },
-            "num_leaves": {
-                "domain": tune.lograndint(lower=4, upper=3276),
-                "init_value": 3276,
-                "low_cost_init_value": 4,
-            },
-        }
-
-
-def custom_metric(
-    X_val,
-    y_val,
-    estimator,
-    labels,
-    X_train,
-    y_train,
-    weight_val=None,
-    weight_train=None,
-    config=None,
-    groups_val=None,
-    groups_train=None,
-):
-    from sklearn.metrics import log_loss
-    import time
-
-    start = time.time()
-    y_pred = estimator.predict_proba(X_val)
-    pred_time = (time.time() - start) / len(X_val)
-    val_loss = log_loss(y_val, y_pred, labels=labels, sample_weight=weight_val)
-    y_pred = estimator.predict_proba(X_train)
-    train_loss = log_loss(y_train, y_pred, labels=labels, sample_weight=weight_train)
-    alpha = 0.5
-    return val_loss * (1 + alpha) - alpha * train_loss, {
-        "val_loss": val_loss,
-        "train_loss": train_loss,
-        "pred_time": pred_time,
-    }
-
-"""
-
-learner_path = customize_learner(learner_code=learner_code)
-from flaml.spark.mylearner import *
+        skip_my_learner = False
+    except ImportError:
+        skip_my_learner = True
+else:
+    skip_my_learner = True
 
 
 class TestMultiClass(unittest.TestCase):
     @unittest.skipIf(skip_spark, "Spark is not installed. Skip all spark tests.")
+    @unittest.skipIf(
+        skip_my_learner,
+        "Please run pytest in the root directory of FLAML, i.e., the directory that contains the setup.py file",
+    )
     def test_custom_learner(self):
         automl = AutoML()
         automl.add_learner(learner_name="RGF", learner_class=MyRegularizedGreedyForest)
@@ -169,6 +65,10 @@ class TestMultiClass(unittest.TestCase):
         automl.fit(X_train=X_train, y_train=y_train, **settings)
 
     @unittest.skipIf(skip_spark, "Spark is not installed. Skip all spark tests.")
+    @unittest.skipIf(
+        skip_my_learner,
+        "Please run pytest in the root directory of FLAML, i.e., the directory that contains the setup.py file",
+    )
     def test_custom_metric(self):
         df, y = load_iris(return_X_y=True, as_frame=True)
         df["label"] = y
@@ -384,6 +284,10 @@ class TestMultiClass(unittest.TestCase):
         print(automl_experiment.best_estimator)
 
     @unittest.skipIf(skip_spark, "Spark is not installed. Skip all spark tests.")
+    @unittest.skipIf(
+        skip_my_learner,
+        "Please run pytest in the root directory of FLAML, i.e., the directory that contains the setup.py file",
+    )
     def _test_memory_limit(self):
         automl_experiment = AutoML()
         automl_experiment.add_learner(
@@ -408,6 +312,10 @@ class TestMultiClass(unittest.TestCase):
         print(automl_experiment.model)
 
     @unittest.skipIf(skip_spark, "Spark is not installed. Skip all spark tests.")
+    @unittest.skipIf(
+        skip_my_learner,
+        "Please run pytest in the root directory of FLAML, i.e., the directory that contains the setup.py file",
+    )
     def test_time_limit(self):
         automl_experiment = AutoML()
         automl_experiment.add_learner(
